Revert "Move association revoke logic to AssociationRevokeProcessor"
This reverts commit 5f07744a75223826f6e7dd13e8b292910ed32b97.

loadAssociationsFromDisk() change makes it call methods on mAssociationRevokeProcessor object in
some cases. mAssociationRevokeProcessor is null at boot-time.

This leads to NullPointerException chain-crashes during boot-time state restoration, which renders
the device unbootable.
